Create README.md
This is a draft of a Mobile application projected to help people to take care of their Mental Health and to get stimulated there are some vantages. 
people that go to the sessions receive vouchers and discounts to go to the gyms, dietitians, personal trainers, doctors, and a team to help with time management.